requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==4.2.15
django-user-agents==0.4.0
djangorestframework==3.15.2
filelock==3.15.4
greenlet==3.0.3
idna==3.8
packaging==24.1
pillow==10.4.0
pyClamd==0.4.0
pyee==11.1.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-whois==0.9.4
requests==2.32.3
requests-file==2.1.0
six==1.16.0
sqlparse==0.5.1
tldextract==5.1.2
typing_extensions==4.12.2
tzdata==2024.1
ua-parser==0.18.0
urllib3==2.2.2
user-agents==2.2.0
